pyproject.toml
[project]
name = "ihearyou"
version = "1.0.0"
description = "Telegram bot with admin panel"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["telegram", "bot", "fastapi", "admin"]

dependencies = [
  "aiogram>=3.22.0",
  "sqlmodel>=0.0.25",
  "fastapi>=0.117.1",
  "uvicorn>=0.37.0",
  "python-dotenv>=1.1.1",
  "sqladmin>=0.21.0",
  "beautifulsoup4>=4.13.5",
  "asyncpg>=0.30.0",
  "pydantic-settings>=2.11.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff", "types-beautifulsoup4>=4.12.0.20250516"]

[project.scripts]
bot = "bot.bot:main"
admin = "admin.admin:app"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
